requirements.txt
streamlit
openai
os
datetime
python-dotenv